New and improved bubble finder!
Working well for PICO data!  Still just a first pass and fairly slow.  Will of course need to be updated and tuned for SBC data once we have bubbles.  Takes path to an event as input, outputs a dictionary with the structure dict[camera][frame][bubble ID][bubble params].  An example of this being used in a notebook can be found in this branch in nb/examples.