Bubble finder set up to take event as input
Takes event from GetEvent as input, gives dictionary of structure cam-->frame-->bubble-->RQs as output.  

Takes various tuning parameters as arguments: 
noise_thresh: threshold that pixels in the difference between a frame and the reference frame must pass, or else they are masked out
rad_thresh: determines whether sig_high or sig_low is used based on the maximum radius found in the previous frame 
sig_high: sigma used for Gaussian blur if the maximum bubble radius in the previous frame is over rad_thresh 
sig_low: sigma used for Gaussian blur if the maximum bubble radius in the previous frame is under rad_thresh
num_pix_in_neighborhood: the radius of a bubble + num_pix_in_neighborhood is the radius inside of which subsequent bubbles will not be found
gof_thresh: goodness of fit threshold that bubble candidates must not exceed in order to declare that the first bubble(s) in an event have been found in this frame

Note that this will not work right now as we have no bubble data and the parameters have not been tuned.  Current parameter defaults are the values that work well for PICO data.  I'll also need to add a key for camera 3 when it's working, or make it so that the module can handle events where certain cameras are missing, whichever is decided.